+++ massrobotics_amr_sender_py/README.md
@@ -1,4 +1,4 @@
-# mrai_sender
+# massrobotics_amr_sender
 
 Configuration-based ROS2 package for sending MassRobotics [AMR Interop Standard messages](https://github.com/MassRobotics-AMR/AMR_Interop_Standard) to complaint receivers.
 
@@ -8,9 +8,9 @@ Make sure `ros2` is installed properly. Clone this repository inside your `src` 
 
 ```bash
 mkdir -p ~/ros2_ws/src && cd ros2_ws/
-git clone https://github.com/inorbit-ai/ros-amr-interop.git ./src
+git clone https://github.com/inorbit-ai/ros_amr_interop.git ./src
 
-colcon build --packages-select mrai_sender
+colcon build --packages-select massrobotics_amr_sender
 ```
 
 ## Running the node
@@ -21,7 +21,7 @@ The node takes the Mass AMR config file path as parameter. If not provided, it i
 # Source the local overlay by running `. install/setup.sh` if
 # using bash or `. install/setup.zsh` if using zsh.
 # Also, install dependencies by running `rosdep install --ignore-src --from-paths src/`
-ros2 run mrai_sender mrai_sender \
+ros2 run massrobotics_amr_sender massrobotics_amr_node \
     --ros-args -p config_file:=/path/to/config.yaml --log-level debug
 ```
 
@@ -30,6 +30,6 @@ ros2 run mrai_sender mrai_sender \
 On you local workspace:
 
 ```bash
-colcon test --packages-select mrai_sender
+colcon test --packages-select massrobotics_amr_sender
 colcon test-result --verbose
 ```
