+++ massrobotics_amr_sender_py/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mrai_sender</name>
+  <name>massrobotics_amr_sender</name>
   <version>0.0.2</version>
   <description>MassRobotics AMR Interop Sender</description>
   <maintainer email="support@inorbit.ai">leandro</maintainer>
