+++ massrobotics_amr_sender_py/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/massrobotics_amr_sender
+[install]
+install-scripts=$base/lib/massrobotics_amr_sender
