--- mrai_sender_py/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/mrai_sender
-[install]
-install-scripts=$base/lib/mrai_sender
